# BaaSPlatform/contact-service/src/main/resources/META-INF/resources/api.yml
openapi: 3.0.1
info:
  title: contact-service API
  version: 1.0.0
  description: |
    Contact Service
servers:
  - url: http://localhost:8074
    description: Generated server url
...
paths:
  /contacts:
    get:
      tags:
        - Contact
      summary: Get the list of contact by user identity
      operationId: getContactList
      parameters:
        - name : userId
          in : query
          description: user id
          required: false
          schema:
            type: string
          example: 123
        - name : searchText
          in : query
          description: search text
          required: false
          schema:
            type: string
          example: abc
        - name : bankCode
          in : query
          description: bank code
          required: false
          schema:
            type: string
          example: UD
        - name: listOrders
          in: query
          description: Results will be sorted and ordered based on the given criteria
          required: false
          schema:
            type: string
          example: firstName-DESC
        - name: pageSize
          in: query
          description: Number of User Accounts returned per page/request
          required: false
          schema:
            type: integer
            default: 10
        - name: pageNumber
          in: query
          description: Get User Accounts in a specific page number
          required: false
          schema:
            type: integer
            default: 1
      responses:
        503:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        200:
          description: Retrieve data succesfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryResponse'
        404:
          description: Data Not Found
          content:
            application/json:
              schema:
                type: string
        400:
          description: Input data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
      security:
        - Bearer: [ ]
    post:
      tags:
        - Contact
      summary: Create a new contact for given user identity
      operationId: createContact
      requestBody:
        description: Contact data model
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeneficiaryRecord'
      responses:
          503:
            description: Internal server error
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          201:
            description: Contact data model created succesfully
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/LookupBeneficiaryResponse'
          404:
            description: Data Not Found
            content:
              application/json:
                schema:
                  type: string
          400:
            description: Input data is invalid
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          401:
            description: Unauthorized
            content:
              application/json:
                schema:
                  type: string
      security:
        - Bearer: [ ]
  /contacts/{id}:
    delete:
      tags:
        - Contact
      summary: Delete the contact using contact id
      operationId: deleteContactById
      parameters:
        - name: id
          in: path
          description: contact id
          required: true
          schema:
            type: string
          example: 123
      responses:
        503:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        204:
          description: Contact deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryResponse'
        404:
          description: Data Not Found
          content:
            application/json:
              schema:
                type: string
        400:
          description: Input data is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - Bearer: [ ]
...
components:
  schemas:
    LookupBeneficiaryResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Beneficiary'
    BeneficiaryResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Beneficiary'
        paging:
          $ref: '#/components/schemas/PagingInformation'
...
    PagingInformation:
      type: object
      properties:
        totalRecords:
          type: integer
        pageSize:
          type: integer
        pageNumber:
          type: integer
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ResponseStatus'
    ResponseStatus:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    Beneficiary:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        paymentReference:
          type: string
        accountNumber:
          type: string
        bankCode:
          type: string
        accountType:
          type: string
        status:
          type: string
        identification:
          type: string
        type:
            type: string
        qrDetails:
          readOnly: true
          $ref: '#/components/schemas/QrDetails'            
    QrDetails:
      type: object
      properties:
        qrType:
          type: string
          enum:
            - static
            - dynamic
        amount:
          type: string
        messageId:
          type: string
        acquirerId:
          type: string
        acquirerName:
          type: string
        category:
          type: string
          enum:
            - POS
            - P2P
            - DFLT
        acceptedSource:
          type: array
          items:
            type: string
        merchantCode:
          type: string
        qrRawData:
          type: string
        originalMessageId:
          type: string 
          example: 20230301FNXSMYNB520HQR02000043 
...
    BeneficiaryRecord:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        serviceCode:
          type: string
          default: ""
        subServiceCode:
          type: string
          default: ""
        displayName:
          type: string
          default: ""
          x-constraints: "@Xss"
        paymentReference:
          type: string
          default: ""
          x-constraints: "@Xss"
        mobileNumber:
          type: string
          default: ""
        accountNumber:
          type: string
          default: ""
          x-constraints: "@Xss"
        branchCode:
          type: string
          default: ""
        bankCode:
          type: string
          default: ""
        city:
          type: string
          default: ""
        state:
          type: string
          default: ""
        postCode:
          type: string
          default: ""
        address:
          type: string
          default: ""
        verificationStatus:
          type: string
          default: ""
        verificationAt:
          type: string
          default: ""
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer